up some fixes

+++ lib/libft/src/lib/ft_strrchr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strrchr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fporto <fporto@student.42.fr>              +#+  +:+       +#+        */
+/*   By: fheaton- <fheaton-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/08 16:55:11 by fporto            #+#    #+#             */
-/*   Updated: 2022/03/09 16:30:42 by fporto           ###   ########.fr       */
+/*   Updated: 2025/08/27 20:45:44 by fheaton-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+	int		i;
 	size_t	len;
 
 	i = 0;
 	len = ft_strlen(s);
 	if (c == '\0')
 		return ((char *)s + len);
-	i = len - 1;
+	i = (int)len - 1;
 	while (i >= 0)
 	{
 		if (s[i] == (unsigned char)c)
